+++ app/src/main/java/com/vanpra/composematerialdialogdemos/demos/BasicDialog.kt
@@ -1,9 +1,10 @@
 package com.vanpra.composematerialdialogdemos.demos
 
+import android.util.Log
+import android.util.Patterns
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.LocationOn
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.res.stringResource
 import com.vanpra.composematerialdialogdemos.DialogAndShowButton
 import com.vanpra.composematerialdialogdemos.R
 
@@ -47,12 +48,32 @@ fun BasicDialogDemo() {
         }
     }
 
-    DialogAndShowButton(buttonText = "Basic Dialog With input") {
+    DialogAndShowButton(buttonText = "Basic Input Dialog") {
         title(res = R.string.input_dialog_title)
         input(label = "Name", hint = "Jon Smith")
         buttons {
             negativeButton("Cancel")
             positiveButton("Ok")
         }
+    }
+
+    DialogAndShowButton(buttonText = "Input Dialog with input validation") {
+        title("Please enter your email")
+        input(
+            label = "Email",
+            hint = "hello@example.com",
+            errorMessage = "Invalid email",
+            isTextValid = {
+                Patterns.EMAIL_ADDRESS.matcher(it).matches() && it.isNotEmpty()
+            }
+        ) {
+            Log.d("SELECTION:", it)
+        }
+
+        buttons {
+            negativeButton("Cancel")
+            positiveButton("Ok")
+        }
+
     }
 }