+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -2,7 +2,7 @@ object Dependencies {
     const val material = "com.google.android.material:material:1.2.1"
 
     object ComposeMaterialDialogs {
-        const val version = "0.2.10"
+        const val version = "0.2.11"
 
         const val core = "com.vanpra.compose-material-dialogs:core:$version"
         const val datetime = "com.vanpra.compose-material-dialogs:datetime:$version"